move cache module to data module
