+++ libraries/data/src/main/AndroidManifest.xml
