+++ libraries/data/src/main/java/com/zizohanto/android/tobuy/data/mappers/ShoppingListWithProductsCacheModelMapper.kt
@@ -1,9 +1,9 @@
-package com.zizohanto.android.tobuy.cache.mappers
+package com.zizohanto.android.tobuy.data.mappers
 
-import com.zizohanto.android.tobuy.cache.mappers.base.CacheModelMapper
-import com.zizohanto.android.tobuy.cache.models.ProductCacheModel
-import com.zizohanto.android.tobuy.cache.models.ShoppingListCacheModel
-import com.zizohanto.android.tobuy.cache.models.ShoppingListWithProductsCacheModel
+import com.zizohanto.android.tobuy.data.mappers.base.CacheModelMapper
+import com.zizohanto.android.tobuy.data.models.ProductCacheModel
+import com.zizohanto.android.tobuy.data.models.ShoppingListCacheModel
+import com.zizohanto.android.tobuy.data.models.ShoppingListWithProductsCacheModel
 import com.zizohanto.android.tobuy.data.models.ShoppingListWithProductsEntity
 import javax.inject.Inject
